Zsolt7 ChallengeService timeManagement

+++ src/main/java/challenger/mink/challenges/ChallengeService.java
@@ -1,5 +1,6 @@
 package challenger.mink.challenges;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.NoSuchElementException;
 import lombok.RequiredArgsConstructor;
@@ -15,9 +16,11 @@ public class ChallengeService {
   }
 
   public void addChallenge(Challenge challenge) {
-    challenge.setStartDate(challenge.getStartDate().plusDays(1));
-    challenge.setEndDate(challenge.getEndDate().plusDays(1));
-    challengeRepository.save(challenge);
+    if (challenge.getStartDate().isAfter(LocalDate.now()) && challenge.getStartDate().isBefore(challenge.getEndDate())) {
+      challenge.setStartDate(challenge.getStartDate().plusDays(1));
+      challenge.setEndDate(challenge.getEndDate().plusDays(1));
+      challengeRepository.save(challenge);
+    }
   }
 
   public Challenge getChallengeById(long id) throws NoSuchChallengeMinkCeption {
@@ -27,6 +30,7 @@ public class ChallengeService {
   public void deleteChallenge(long challengeId) {
     Challenge challenge = challengeRepository.findById(challengeId).orElseThrow(
         NoSuchElementException::new);
+    if(LocalDate.now().isBefore(challenge.getStartDate()))
     challengeRepository.delete(challenge);
   }
 }
